+++ src/main/kotlin/ru/mipt/npm/gradle/KScienceJVMPlugin.kt
@@ -13,5 +13,4 @@ open class KScienceJVMPlugin : Plugin<Project> {
         }
         plugins.apply(KScienceCommonPlugin::class)
     }
-
 }